refactor: split settings and model management from geosam runtime | make downloading models works
